Fixing broken post of new kanji to tool.
It used to be when posting a new kanji to the database, it would
try to update a null kanji item instead of creating a new one (it makes
sure that kanji doesn't already exist in database). After this commit,
a new kanji model will properly be saved to database.