Fixing bug so user's kanji are updateable
The put method verifies that the kanji being updated belongs to the user
that is logged-in. The comparison, however, was between an object and a
string, and so the code always decided that the kanji being updated did
not belong to the user trying to update it.

JSON.stringifying both the req's user id and the kanji user id before
comparing. This code will require further cleanup in the future, but
works for right now.